Moved encoding/decoding of item data into SyncItem::changeDataEncoding. Encoding used to be done by the SyncMLBuilder, decoding by the SyncManager.
The "encryption" config property became obsolete, it was replaced by
"encoding = des;b64". However, due to a lack of actual DES support this
mode is untested.

Outgoing data is encoded automatically according to the source configuration
by SyncManager::getItem(), a wrapper around the various SyncItem methods
which return a new item to send. Incoming data is decoded inside
SyncManager::processSyncItem() before passing it to the SyncSource. So in this
mode the change is transparent to clients.

In addition to letting the client library do everything, the client can also
decide to do the encoding/decoding itself and together with a suitable
server add encodings not supported by the core library.


git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@10661 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10