ArrayList modified to improve the performance.
The method size() return a value that is incremented and decremented everytime there's an add or a remove.

The method add is also improved when there is an add in the last element (a sort of pushback)

Removed the remove method now there's only the method removeElementAt(index)

in SyncSourceReport in the SyncSourceReport::addItem method was removed a for cycle that check if there is a element clone of the one to add. This for cycle is very negative for the performance and useless.


git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@17837 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10