added CtCap.
we pass to the formatter an ArrayList of CtCap to handle different capabilities.

in SyncSourceConfig there's a new method called addCtCap to add an element to the ArrayList of CtCap
this method has 4 parameters CtType VerCt an arraylist of properties and an int with 3 status to handle the FieldLevel param (FLEVEL_DISABLED, FLEVEL_ENABLED, FLEVEL_UNDEFINED)
CtCype VerCt and FieldLevel are optional. If one of this parameters is not set the method use the default one setted in the configuration

getCtCaps returns a reference to the ArrayList of CtCaps


git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@22046 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10