Creates two abstract base classes for sync source and sync manager configs. Access to these properties inside the SyncManager gets simpler because the additional redirection via DeviceConfig and AccessConfig gets removed.
This patch is transparent to clients because the traditional classes
remain mostly unchanged: SyncManagerConfig "is an"
AbstractSyncConfig and implements its methods via proxies into
the DeviceConfig and AccessConfig that it "has".

At a later point the traditional classes could be merged and/or moved
to "common/client" to make it clear that they are neither required nor
used by the core code. Merging SyncManagerConfig, DeviceConfig and
AccessConfig would allow to remove the proxy code.

The patch cleans up SyncManagerConfig a bit by moving simple get
methods into the header file. While doing that I noticed that
"isDirty()" seems to be incomplete - is it obsolete?

The criteria for deciding which functions belong into the
base classes were:
- anything which is required by the core code
- no set*() methods unless required
- no redundant functions (in particular if they unnecessarily
  impose implementation limitations, like the "array of pointers"
  in SyncManagerConfig::getSyncSourceConfigs())
- no utility functions which depend on a specific implementation

The Doxygen documentation is as good (or bad) as it was before because
the base classes were created by cut-and-paste of the existing
headers. The free-form comments in DeviceConfig.h and AccessConfig.h
were converted into method documentation for AbstractSyncConfig.h.

The recently added getCtCaps() method lacked "const" attributes. I
added this in addition to the non-const version because it is required
for implementations which really only have read-only CtCaps and
because having "const" is consistent with the rest of the get*()
methods. The drawback is that DataStore unnecessarily requires
non-const CtCaps: trying to fix that made the patch unnecessarily
large, so I went for a type-cast hack instead.


git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@22321 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10