destructor now frees resource via dispose()
This change was necessary because on Linux with shared objects
the destructor was called twice, once when cleaning up the main
executable and again when unloading the shared object. Instead
of investigating this it was easier to use dispose() (which
resets the pointer) instead of the "delete &" (which left
a dangling pointer, leading to a double free).

git-svn-id: https://core.forge.funambol.org/svn/core/top-level/trunk/3x/client-api/native@26931 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10