CurlTransportAgent: avoid troublesome Expect: 100-continue header
As pointed out by SyncEvolution users [1], some proxies do not support
the "Expect: 100-continue" behavior which libcurl uses by default.
The expected behavior is that the header is forwarded to the server,
the server checks the POST header and then asks the client for
the actual data. When a proxy doesn't know the option, then it
(correctly) rejects the message without forwarding it.

The optimization with 100-continue is meant to avoid transmitting data
unnecessarily when the server would reject it for other reasons. This
should never happen with SyncML, or at least only for the first
message, which is small. Therefore this patch disables the header
unconditionally.

[1] https://sourceforge.net/tracker/index.php?func=detail&aid=2101015&group_id=146288&atid=764733

git-svn-id: https://core.forge.funambol.org/svn/core/trunk/funambol/client-api/native@29250 e8e8ed6c-164c-0410-afcf-9e9a7c7d8c10