avoid segfault after failed authentication with Memotoo server
The Memotoo server replies to invalid credentials with 407 status
*without* sending a challenge. In one of the two code branches which
handle failed authentication a check for availability of the
challenge was missing, leading to an access via a NULL pointer; the
other branch already had the check.

Adding the check leads to the desired result: first try with md5 auth,
then plain auth, then abort with "Client not authenticated".