Multiple listeners.
Listener mechanism changed to support multiple listener registered for an event.
Each listener has now a name, with default to empty.
Setting twice a listener with the same name removes the previous and set the new
one. Not passing the name keeps the old behavior, so there is no impact on
existing clients.



git-svn-id: https://client-sdk.forge.funambol.org/svn/client-sdk/trunk/cpp-sdk@87 3966be30-755a-0410-815e-dd79f492ac3c