SyncManager::endSync(): if a network problem ends the session prematurely, then don't update anchors
If the last message to the server cannot be sent, then the next sync
would be done in "slow" mode because the client has updated its anchor
but the server hasn't. This patch fixes by checking for a network
error during endSync().

Client and server must both notice that the last message
exchange failed, otherwise there will still be a mismatch
between anchors (client noticed problem, server didn't: anchor
on client not updated, but updated on server).